logiccraft_backend/requirements.txt
django>=4.2
 djangorestframework
 redis
 django-cors-headers
 requests
 celery
 django-environ

chromadb
sentence-transformers
